app/requirements.txt
aiohttp==3.7.4.post0; python_version >= "3.6"
anyio==3.6.2; python_full_version >= "3.6.2" and python_version >= "3.7"
async-timeout==3.0.1; python_version >= "3.6" and python_full_version >= "3.5.3"
asyncpg==0.27.0; python_version >= "3.7" and python_full_version >= "3.7.0"
attrs==22.2.0; python_version >= "3.7"
backoff==2.2.1; python_version >= "3.7" and python_version < "4.0"
botocore==1.29.64; python_version >= "3.7"
certifi==2022.12.7; python_version >= "3.7" and python_version < "4"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.5.0"
charset-normalizer==3.0.1; python_version >= "3.7" and python_version < "4"
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.7.0" and platform_system == "Windows"
databases==0.7.0; python_version >= "3.7"
exceptiongroup==1.1.0; python_version < "3.11" and python_version >= "3.7"
fastapi==0.89.1; python_version >= "3.7"
gql==3.4.0
graphql-core==3.2.3; python_version >= "3.6" and python_version < "4"
greenlet==2.0.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and python_full_version >= "3.5.0"
h11==0.14.0; python_version >= "3.7"
idna==3.4; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.2"
iniconfig==2.0.0; python_version >= "3.7"
jmespath==1.0.1; python_version >= "3.7"
multidict==6.0.4; python_version >= "3.7"
packaging==23.0; python_version >= "3.7"
pluggy==1.0.0; python_version >= "3.7"
psycopg2-binary==2.9.5; python_version >= "3.6"
pydantic==1.10.4; python_version >= "3.7"
pytest==7.2.1; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.3.0"
python-dotenv==0.21.1; python_version >= "3.7"
requests-toolbelt==0.10.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.28.2; python_version >= "3.7" and python_version < "4" and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0")
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.3.0"
sniffio==1.3.0; python_full_version >= "3.6.2" and python_version >= "3.7"
sqlalchemy==1.4.46; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
starlette==0.22.0; python_version >= "3.7"
tomli==2.0.1; python_version < "3.11" and python_version >= "3.7"
typer==0.7.0; python_version >= "3.6"
typing-extensions==4.4.0; python_version >= "3.7"
urllib3==1.26.14; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version >= "3.7" and python_version < "4"
uvicorn==0.20.0; python_version >= "3.7"
websockets==10.4; python_version >= "3.7"
yarl==1.8.2; python_version >= "3.7"
